+++ src/components/sbb-alert/sbb-alert.events.ts
@@ -3,7 +3,7 @@
  * See stencil.config.ts in the root directory.
  */
 export default {
-  didPresent: 'sbb-alert_did-present',
-  dismissalRequested: 'sbb-alert_dismissal-requested',
-  willPresent: 'sbb-alert_will-present',
+  didPresent: 'did-present',
+  dismissalRequested: 'dismissal-requested',
+  willPresent: 'will-present',
 };