+++ src/components/sbb-card-product/sbb-card-product.events.ts
@@ -1,3 +1,3 @@
 export default {
-  click: 'sbb-card-product_click',
+  click: 'click',
 };