+++ src/global/helpers/testing/wait-for-stable-position.ts
@@ -2,7 +2,7 @@ import { waitFor } from '@storybook/testing-library';
 
 export async function waitForStablePosition(
   elementCallback: () => HTMLElement,
-  stableDurationMs = 400
+  stableDurationMs = 2000
 ): Promise<void> {
   const element = await waitFor(elementCallback);
   const getPositionString = (): string => JSON.stringify(element.getBoundingClientRect());